better training with better cropping
