+++ dataloader.py
@@ -3,6 +3,7 @@
 
 import torch 
 from torch.utils.data import Dataset, DataLoader
+from pytorch_lightning import LightningDataModule
 import numpy as np
 from PIL import Image
 import imageio
@@ -16,38 +17,27 @@ def image_to_tensor(image):
         tensor = torch.from_numpy(image).permute(2,0,1).float()
     return tensor
 
-def sample_random_coordinates(N, height, width, alpha=None, prop=0.0): 
+def sample_random_coordinates(N, height, width, cropping=False): 
     """Two [N,] torch tensors representing random coordinates.
 
     Args:
         N: int representing number of coordinates to sample
         height: the maximum height value (exclusive)
         width: maximum width value (exclusive)
-        alpha: alpha channel of an image, used to weight coordinates to sample.
-               No weighting if None.
-        prop: minimum proportion of coordinates that have to 
-                    have alpha values > 0.
+        cropping: samples from center of image if true, otherwise
+                  uniformly samples whole image.
     Returns:
         xs: [N,] torch tensor of random ints [0,width)
         ys: [N,] torch tensor of random ints [0,height)
     """
-    if alpha == None or prop == 0:
+    if cropping:
+        edge_width = width // 4
+        edge_height = height // 4
+        xs = torch.randint(edge_width, width - edge_width, size=(N,))
+        ys = torch.randint(edge_height, height - edge_height, size=(N,))
+    else:
         xs = torch.randint(0, width, size=(N,))
         ys = torch.randint(0, height, size=(N,))
-    else: 
-        num_in_alpha = int(N * prop)
-        num_random = N - num_in_alpha
-        # pseudo science to pick an initial number of samples
-        num_samples = int(N * (1 / (1 - prop)))
-        xs = torch.randint(0, width, size=(num_samples,))
-        ys = torch.randint(0, height, size=(num_samples,))
-        # keep the ones that have alpha > 0.
-        valid = alpha[ys, xs]
-        xs = xs[valid]; ys = ys[valid]
-        if xs.shape[0] > num_in_alpha:
-            xs = xs[:num_in_alpha]; ys = ys[:num_in_alpha]
-        xs = torch.cat((xs, torch.randint(0, width, size=(num_random,))))
-        ys = torch.cat((ys, torch.randint(0, height, size=(num_random,))))
     return xs, ys
 
 def get_rays(H, W, focal, c2w):
@@ -59,40 +49,36 @@ def get_rays(H, W, focal, c2w):
     rays_o = torch.broadcast_to(c2w[:3, -1], rays_d.shape)
     return rays_o, rays_d
 
-def convert_to_ndc_rays(o_rays, d_rays, focal, width, height, near=1.0): 
-    """This is only for FRONT-FACING scenes.
+class SyntheticDataModule(LightningDataModule):
+    def __init__(self, base_dir, num_rays, cropping_epochs, num_workers=8):
+        self.num_rays = num_rays
+        self.base_dir = base_dir
+        self.cropping_epochs = cropping_epochs
+        self.num_workers = num_workers
+        self.crop_train_ds = SyntheticDataset(base_dir, 'train', num_rays,
+                                              cropping=True)
+        self.train_ds = SyntheticDataset(base_dir, 'train', num_rays,
+                                         cropping=False)
+        self.val_ds = SyntheticDataset(base_dir, 'val', num_rays,
+                                        cropping=False)
+        
+    def train_dataloader(self):
+        if self.trainer.current_epoch < self.cropping_epochs:
+            return DataLoader(dataset=self.crop_train_ds, batch_size=1,
+                              shuffle=True, num_workers=self.num_workers)
+        else:
+            return DataLoader(dataset=self.train_ds, batch_size=1, shuffle=True,
+                              num_workers=self.num_workers)
 
-    Args:
-        o_rays: [H x W x 3] representing ray origin.
-        d_rays: [H x W x 3] representing ray direction.
-        focal: focal length.
-        width: the maximum width 
-        height: the maximum height
-        near: the near depth bound (i.e. 1)
-    """
-    # shift o to the ray’s intersection with the near plane at z = −n
-    t_near  = - (near + o_rays[:,:,2] ) / d_rays[:,:,2] 
-    o_rays = o_rays + t_near[...,None] * d_rays
-
-    ox, oy, oz = o_rays[:,:,0], o_rays[:,:,1], o_rays[:,:,2] 
-    dx, dy, dz = d_rays[:,:,0], d_rays[:,:,1], d_rays[:,:,2] 
-    
-    ox_new =  -1.0 * focal / (width / 2) * (ox / oz)
-    oy_new =  -1.0 * focal / (height / 2) * (oy / oz)
-    oz_new = 1.0 + (2 * near) / oz
-    dx_new =  -1.0 * focal / (width / 2) * ((dx / dz) - (ox / oz))
-    dy_new =  -1.0 * focal / (height / 2) * ((dy / dz) - (oy / oz))
-    dz_new = (- 2 * near) / oz
-    
-    o_ndc_rays = torch.stack([ox_new, oy_new, oz_new], axis=-1)
-    d_ndc_rays = torch.stack([dx_new, dy_new, dz_new], axis=-1)
-    d_ndc_rays = d_ndc_rays / torch.linalg.norm(d_ndc_rays, dim=-1,keepdim=True)
-    return o_ndc_rays, d_ndc_rays
+    def val_dataloader(self): 
+        return DataLoader(dataset=self.val_ds, batch_size=1,
+                          shuffle=False, num_workers=self.num_workers)
+        
 
 class SyntheticDataset(Dataset):
     """Dataset for Synthetic images."""
 
-    def __init__(self, base_dir, tvt, num_rays, prop):
+    def __init__(self, base_dir, tvt, num_rays, cropping=False):
         """Constructor.
 
         Args:
@@ -100,7 +86,7 @@ class SyntheticDataset(Dataset):
             tvt: str of either "train", "val", or "test".
             num_rays: number of rays to sample per batch / image.
         """ 
-        self._setup(base_dir, tvt, num_rays, prop)
+        self._setup(base_dir, tvt, num_rays, cropping)
         file = open(Path(self.base_dir, f'transforms_{tvt}.json'))
         self.data = json.load(file)
         self.frames = self.data['frames']
@@ -109,12 +95,12 @@ class SyntheticDataset(Dataset):
         self._preprocess()
         file.close()
 
-    def _setup(self, base_dir, tvt, num_rays, prop): 
+    def _setup(self, base_dir, tvt, num_rays, cropping): 
         # we know synthetic images are 800x800.
         self.H = 800
         self.W = 800
-        self.prop = prop
         self.tvt = tvt
+        self.cropping = cropping
         self.num_rays = num_rays
         self.base_dir = base_dir
 
@@ -133,8 +119,7 @@ class SyntheticDataset(Dataset):
         cam_to_world = torch.Tensor(frame['transform_matrix']) 
         o_rays, d_rays = get_rays(self.H, self.W, self.focal, cam_to_world)  # [HxWx3]
         image = (torch.Tensor(imageio.imread(frame['file_path'], pilmode="RGBA")) / 255.0).float()  # [HxWx3]
-        xs, ys = sample_random_coordinates(self.num_rays, self.H, self.W, 
-                                           image[:,:,3].bool(), prop=self.prop) # (N,), (N,)
+        xs, ys = sample_random_coordinates(self.num_rays, self.H, self.W, cropping=self.cropping) # (N,), (N,)
         rgb = image[ys,xs,:3]
         origin = o_rays[ys, xs, :]
         direction = d_rays[ys, xs, :]
@@ -145,8 +130,8 @@ class SyntheticDataset(Dataset):
             return {'origin': origin, 'direc': direction, 'rgb': rgb, 'xs': xs, 'ys': ys,
                     'all_origin': o_rays, 'all_direc': d_rays}
 
-def getSyntheticDataloader(base_dir, tvt, num_rays, prop=0.0, num_workers=8, shuffle=True): 
-    dataset = SyntheticDataset(base_dir, tvt, num_rays, prop=prop)
+def getSyntheticDataloader(base_dir, tvt, num_rays, cropping=False, num_workers=8, shuffle=True): 
+    dataset = SyntheticDataset(base_dir, tvt, num_rays, cropping=cropping)
     return DataLoader(dataset=dataset, batch_size=1, shuffle=shuffle, num_workers=num_workers)
 
 class PhotoDataset(Dataset):
